pyproject.toml
[project]
name = "stock-dashboard"
version = "0.1.0"
description = "미국 배당주 데이터와 S&P 500 데이터를 처리하는 데이터 파이프라인"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "yfinance>=0.2.18",
    "pandas>=1.5.0",
    "numpy>=1.24.0",
    "requests>=2.28.0",
    "lxml>=4.9.0",
    "html5lib>=1.1",
    "beautifulsoup4>=4.11.0",
    "finance-datareader>=0.9.50",
    "pyarrow>=10.0.0",
    "google-cloud-storage>=2.0.0",
    "deltalake>=0.15.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.1",
    "pytest-cov>=5.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=3.0.0",
    "black>=22.0.0",
    "isort>=5.10.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]
test = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "pytest-cov>=3.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
